requirements.txt
pyspark==3.5.1
delta-spark==3.1.0
numpy==1.26.4
pandas==2.2.1
httpx==0.27.0
azure-storage-blob==12.19.1
azure-core==1.30.1
